Fix compiler warnings (STR 2988), porting from branch-1.3.
This commit is the accumulated patch introduced in branch 1.3
in svn r 11094, 11095, and 11096.


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3-porting@11097 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/vsnprintf.c
@@ -128,7 +128,7 @@ int fl_vsnprintf(char* buffer, size_t bufsize, const char* format, va_list ap) {
 	case 'e' :
 	case 'f' :
 	case 'g' :
-	  if ((width + 2) > sizeof(temp)) break;
+	  if ((width + 2) > (int)sizeof(temp)) break;
 
 	  sprintf(temp, tformat, va_arg(ap, double));
 
@@ -153,7 +153,7 @@ int fl_vsnprintf(char* buffer, size_t bufsize, const char* format, va_list ap) {
 	case 'o' :
 	case 'u' :
 	case 'x' :
-	  if ((width + 2) > sizeof(temp)) break;
+	  if ((width + 2) > (int)sizeof(temp)) break;
 
 #ifdef HAVE_LONG_LONG
 	  if (size == 'L')
@@ -179,7 +179,7 @@ int fl_vsnprintf(char* buffer, size_t bufsize, const char* format, va_list ap) {
 	  break;
 	    
 	case 'p' : /* Pointer value */
-	  if ((width + 2) > sizeof(temp)) break;
+	  if ((width + 2) > (int)sizeof(temp)) break;
 
 	  sprintf(temp, tformat, va_arg(ap, void *));
